requirements.txt
alpaca-trade-api==3.1.1
lumibot==3.6.24
ib_async==1.0.3
plotly==5.23.0
python-dotenv==0.20.0
timedelta==2020.12.3
# torch==2.2.0
# transformers==4.37.2
yfinance==0.2.41
pip3-autoremove==1.2.2
wheel==0.44.0
pandas==2.2.0
matplotlib==3.8.2
numpy==1.26.4

notebook==7.2.1
black[jupyter]==23.9.1
flake8==6.1.0
ipytest==0.13.3
pre-commit==3.4.0
pyclean==2.7.5
pytest==8.3.2
